% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze2DData.R
\name{analyze2DData}
\alias{analyze2DData}
\title{analyze2DData}
\usage{
analyze2DData(x2)
}
\arguments{
\item{x2}{Results from chi-square test of association}
}
\value{
List containing observed and expected frequenty data for plotting.
}
\description{
\code{analyze2DData} Render frequency and proportion data
}
\seealso{
Other xmar functions: \code{\link{analyzeTrend}},
  \code{\link{bivariate}}, \code{\link{diffTest}},
  \code{\link{multivariate}}, \code{\link{plotBars}},
  \code{\link{plotDiffTest}}, \code{\link{plotLines}},
  \code{\link{plotX2}}, \code{\link{preprocess}},
  \code{\link{univariate}}, \code{\link{x2Test}},
  \code{\link{x2TrendTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
