% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeTrend.R
\name{analyzeTrend}
\alias{analyzeTrend}
\title{analyzeTrend}
\usage{
analyzeTrend(tbl, x, y, z, dim)
}
\arguments{
\item{tbl}{Table containing series data}

\item{x}{Cheracter string indicating the explanatory variable}

\item{y}{Cheracter string indicating the response variable}

\item{z}{Cheracter string indicating the level of the control variable}

\item{dim}{Character string indicating the dimension of the table from which the series data is extracted}
}
\value{
data frame of summary statistics
}
\description{
\code{analyzeTrend} Provides numeric summary of series data.
}
\seealso{
Other xmar functions: \code{\link{analyze2DData}},
  \code{\link{bivariate}}, \code{\link{diffTest}},
  \code{\link{multivariate}}, \code{\link{plotBars}},
  \code{\link{plotDiffTest}}, \code{\link{plotLines}},
  \code{\link{plotX2}}, \code{\link{preprocess}},
  \code{\link{univariate}}, \code{\link{x2Test}},
  \code{\link{x2TrendTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
