% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffTest.R
\name{diffTest}
\alias{diffTest}
\title{diffTest}
\usage{
diffTest(data, alternative = "two.sided", success, conf = 0.95,
  alpha = 0.05)
}
\arguments{
\item{data}{data frame containing data to be analyzed, with the response variable as the first variable.}

\item{alternative}{direction of the alternative hypothesis; "less","greater", or "two.sided"}

\item{success}{Character string indicating which level of the response to consider the "success"}

\item{conf}{confidence level, value between 0 and 1}

\item{alpha}{numeric between 0 and 1, the probability of a type I error}
}
\value{
Data frame containing the results of the z-test
}
\description{
\code{diffTest} Conducts multiple z-tests for a multi-level categorical explanatory
variable on a response variable.
}
\seealso{
Other xmar functions: \code{\link{bivariate}},
  \code{\link{formatData}}, \code{\link{multivariate}},
  \code{\link{plotBars}}, \code{\link{plotDiffTest}},
  \code{\link{plotLines}}, \code{\link{plotX2}},
  \code{\link{preprocess}}, \code{\link{univariate}},
  \code{\link{x2Test}}, \code{\link{x2TrendTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
