% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBars.R
\name{plotBars}
\alias{plotBars}
\title{plotBars}
\usage{
plotBars(data, y, x, title)
}
\arguments{
\item{data}{Data frame containing proportions comprised of two columns: 'target' and 'group'.
'target' represents the proportions for the target variable and 'group' is the name of the grouping variable}

\item{y}{Character string indicating the name of the response variable.}

\item{x}{Character string indicating the name of the explanatory variable.}

\item{title}{Character string indicating the title of the plot}
}
\description{
\code{plotBars} Renders segmented bar plots with value labels representing
proportions each segment represents.
}
\seealso{
Other xmar functions: \code{\link{analyze2DData}},
  \code{\link{analyzeTrend}}, \code{\link{bivariate}},
  \code{\link{diffTest}}, \code{\link{multivariate}},
  \code{\link{plotDiffTest}}, \code{\link{plotLines}},
  \code{\link{plotX2}}, \code{\link{preprocess}},
  \code{\link{univariate}}, \code{\link{x2Test}},
  \code{\link{x2TrendTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
