% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiffTest.R
\name{plotDiffTest}
\alias{plotDiffTest}
\title{plotDiffTest}
\usage{
plotDiffTest(x, y, zAlpha, zScore, lb = -Inf, ub = Inf, mean = 0,
  sd = 1, limits = c(mean - 3 * sd, mean + 3 * sd))
}
\arguments{
\item{x}{= Character string with name of group associated with proportion 1}

\item{y}{= Character string with name of group associated with proportion 2}

\item{lb}{Lower bound of the shaded area. Use \code{-Inf} for a left tail.}

\item{ub}{Upper bound of the shaded area. Use \code{Inf} for a right tail.}

\item{mean}{Mean of the normal distribution}

\item{sd}{Standard deviation of the normal distribution}

\item{limits}{Lower and upper bounds on the x-axis of the area displayed.}

\item{dp}{Dataframe summarizing results of a difference in proportions test}
}
\value{
ggplot object.
}
\description{
\code{plotDiffTest} Renders a plot of the normal distribution plot for
difference in proportions test with critical regions shaded.
}
\seealso{
Other xmar functions: \code{\link{analyze2DData}},
  \code{\link{analyzeTrend}}, \code{\link{bivariate}},
  \code{\link{diffTest}}, \code{\link{multivariate}},
  \code{\link{plotBars}}, \code{\link{plotLines}},
  \code{\link{plotX2}}, \code{\link{preprocess}},
  \code{\link{univariate}}, \code{\link{x2Test}},
  \code{\link{x2TrendTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
