% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLines.R
\name{plotLines}
\alias{plotLines}
\title{plotLines}
\usage{
plotLines(data, y, x, title)
}
\arguments{
\item{data}{Data frame with Opinion, Gender and Year}

\item{y}{Character string indicating the name of the response variable.}

\item{x}{Character string indicating the name of the explanatory variable.}

\item{title}{Character string indicating the title of the plot}
}
\description{
\code{plotLines} Renders a two dimensional line plot of proportions of
opinion by gender
}
\seealso{
Other xmar functions: \code{\link{bivariate}},
  \code{\link{diffTest}}, \code{\link{formatData}},
  \code{\link{multivariate}}, \code{\link{plotBars}},
  \code{\link{plotDiffTest}}, \code{\link{plotX2}},
  \code{\link{preprocess}}, \code{\link{univariate}},
  \code{\link{x2Test}}, \code{\link{x2TrendTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
