% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotX2.R
\name{plotX2}
\alias{plotX2}
\title{plotX2}
\usage{
plotX2(x2)
}
\arguments{
\item{x2}{Dataframe summarizing results of a chi-square test}
}
\description{
\code{plotX2} Renders a plot of the chi-square distribution for a chi-square
test with the upper tail shaded.
}
\seealso{
Other xmar functions: \code{\link{bivariate}},
  \code{\link{diffTest}}, \code{\link{formatData}},
  \code{\link{multivariate}}, \code{\link{plotBars}},
  \code{\link{plotDiffTest}}, \code{\link{plotLines}},
  \code{\link{preprocess}}, \code{\link{univariate}},
  \code{\link{x2Test}}, \code{\link{x2TrendTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
