% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(GSS)
}
\description{
\code{preprocess} Creates additional variables required for analysis.
}
\seealso{
Other xmar functions: \code{\link{analyze2DData}},
  \code{\link{analyzeTrend}}, \code{\link{bivariate}},
  \code{\link{diffTest}}, \code{\link{multivariate}},
  \code{\link{plotBars}}, \code{\link{plotDiffTest}},
  \code{\link{plotLines}}, \code{\link{plotX2}},
  \code{\link{univariate}}, \code{\link{x2Test}},
  \code{\link{x2TrendTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
