% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2Test.R
\name{x2Test}
\alias{x2Test}
\title{x2Test}
\usage{
x2Test(freqTbl, y, x, conf = 0.95, alpha = 0.05)
}
\arguments{
\item{y}{Character string indicating the name of the response variable}

\item{x}{Character string indicating the name of the explanatory variable}

\item{conf}{Level of confidence between 0 and 1}

\item{alpha}{The probability of a type 1 error between 0 and 1}

\item{data}{Data to be analyzed with response variable as first variable.}
}
\description{
\code{x2Test} Conducts a chi-square test of independency on a contingency table
and renders a data frame containing the results and a statement describing
the results
}
\seealso{
Other xmar functions: \code{\link{analyze2DData}},
  \code{\link{analyzeTrend}}, \code{\link{bivariate}},
  \code{\link{diffTest}}, \code{\link{multivariate}},
  \code{\link{plotBars}}, \code{\link{plotDiffTest}},
  \code{\link{plotLines}}, \code{\link{plotX2}},
  \code{\link{preprocess}}, \code{\link{univariate}},
  \code{\link{x2TrendTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
